+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GITHUB-77 Fix missing new column for jobs migration
 - GITHUB-74 Add possibility to define defaults responses
 - Fix MySQL query parameter that needs to be escaped
+- GITHUB-78 Add support for null as database password
 
 # 1.0.0-alpha1