+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix MySQL query parameter that needs to be escaped
 - GITHUB-78 Add support for null as database password
 - GITHUB-84 Fix missing internal job instance
+- GITHUB-79 Remove useless usage of database name in MySQL queries
 
 # 1.0.0-alpha1