+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.0.0-x
+# 1.0.0-alpha2
 
 ## Improvements